+++ Sources/XcodeGenKit/SourceGenerator.swift
@@ -51,6 +51,7 @@ class SourceGenerator {
     func generateSourceFile(targetType: PBXProductType, targetSource: TargetSource, path: Path, buildPhase: TargetSource.BuildPhase? = nil) -> SourceFile {
         let fileReference = fileReferencesByPath[path.string.lowercased()]!
         var settings: [String: Any] = [:]
+        var attributes: [String] = targetSource.attributes
         var chosenBuildPhase: TargetSource.BuildPhase?
 
         let headerVisibility = targetSource.headerVisibility ?? .public
@@ -80,11 +81,15 @@ class SourceGenerator {
         if chosenBuildPhase == .headers {
             if headerVisibility != .project {
                 // Xcode doesn't write the default of project
-                settings["ATTRIBUTES"] = [headerVisibility.settingName]
+                attributes.append(headerVisibility.settingName)
             }
         }
         if chosenBuildPhase == .sources && targetSource.compilerFlags.count > 0 {
             settings["COMPILER_FLAGS"] = targetSource.compilerFlags.joined(separator: " ")
+        }
+
+        if !attributes.isEmpty {
+            settings["ATTRIBUTES"] = attributes
         }
 
         let buildFile = PBXBuildFile(file: fileReference, settings: settings.isEmpty ? nil : settings)